Add Aux sKey Postgres repository

+++ short_link/internal/repository/db/migration/000001_init_schema.up.sql
@@ -39,11 +39,11 @@ CREATE TABLE recycle_link (
 );
 
 CREATE TABLE aux_s_key (
-                           N int,
-                           "End" int,
-                           Step int,
-                           A0 int,
-                           N0 int
+                           N bigint,
+                           "End" bigint,
+                           Step bigint,
+                           A0 bigint,
+                           N0 bigint
 );
 
 CREATE INDEX ON users (id);
